requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.11.0
pytz==2019.3
sqlparse==0.3.0
